[Refactor] const 폴더로 카테고리 리스트 이동 #261
